+++ src/graph/controller/item.js
@@ -5,6 +5,7 @@ const NODE = 'node';
 const EDGE = 'edge';
 const CFG_PREFIX = 'default';
 const MAPPER_SUFFIX = 'Mapper';
+const STATE_SUFFIX = 'stateStyles';
 const hasOwnProperty = Object.hasOwnProperty;
 
 class ItemController {
@@ -16,14 +17,14 @@ class ItemController {
     const parent = graph.get(type + 'Group') || graph.get('group');
     const upperType = Util.upperFirst(type);
     let item;
-    let styles = graph.get(type + 'Style') || {};
+    let styles = graph.get(type + Util.upperFirst(STATE_SUFFIX)) || {};
     const defaultModel = graph.get(CFG_PREFIX + upperType);
     const mapper = graph.get(type + MAPPER_SUFFIX);
     if (mapper) {
       const mappedModel = mapper(model);
-      if (mappedModel.styles) {
-        styles = mappedModel.styles;
-        delete mappedModel.styles;
+      if (mappedModel[STATE_SUFFIX]) {
+        styles = mappedModel[STATE_SUFFIX];
+        delete mappedModel[STATE_SUFFIX];
       }
       Util.each(mappedModel, (val, cfg) => {
         model[cfg] = val;