+++ src/util/index.js
@@ -8,5 +8,7 @@ const MathUtil = require('./math');
 const PathUtil = require('./path');
 const BaseUtil = require('./base');
 const GraphicUtil = require('./graphic');
-BaseUtil.deepMix(Util, BaseUtil, GraphicUtil, PathUtil, MathUtil);
+const GroupUtil = require('./groupData');
+const TextUtil = require('./text');
+BaseUtil.deepMix(Util, BaseUtil, GraphicUtil, PathUtil, MathUtil, GroupUtil, TextUtil);
 module.exports = Util;